[IMP] web: make header in day/week calendar clickable
Add the option to create all day events by clicking the date header, similar to
google calendar. This was previously doable by clicking the unlabeled 'all day'
row below the header.

The calendar tour originally contained a step which clicked the header. This
would now open a modal causing the tour to fail with:
Element has been found.
BUT: It is not allowed to do action on an element that's below a modal.
-> changed the step to hover to avoid the issue, as the click was not necessary

The code removed from hr_homeworking_calendar is no longer necessary, because
`handleWorkLocationClick` is also called in `onDateClick` which is now called
in the parent model.

Changed the row size to better match the Google Calendar look. This change
required some edits to the tests.
First, this changed by how many pixels we scroll to reach a specific time.
Second, when performing drag and drop, the drop would place the event slightly
too low, resulting in an incorrect time. For example when an event starting at
8:00 would be dragged to a different day, it could result in it being placed
into the 8:15 slot. Slightly reducing the Y offset of the drop location has
solved the issue.

Task-5421331
